+++ R/distributions.R
@@ -10,6 +10,12 @@
 #' @param detQ,detU,detV Pre-computed log-determinants
 #' @param log Should the log-density be returned?
 #' @param n Number of deviates to produce.
+#' @param rate the Exponential rate parameter
+#' @param a,b Bounds for the truncated exponential distribution.
+#' @param q A quantile
+#' @param p A probability
+#' @param lower.tail Should lower tail probabilities be returned (default) or upper?
+#' @param log.p Should the log be returned?
 #' @param alpha the Dirichlet parameter vector or matrix.
 #' @details
 #'   \code{dmvnorm_diff(x, y, mu, Q)} is equivalent to (but usually twice as fast as)
@@ -161,4 +167,46 @@ rdirich <- function(n, alpha) {
   gam <- stats::rgamma(length(alpha), alpha, rate = 1)
   dim(gam) <- dim(alpha)
   gam / rowSums(gam)
+}
+
+#' @rdname distributions
+#' @export
+rtruncexp <- function(n, rate = 1, a = 0, b = Inf) {
+  stopifnot(b > a)
+  u <- stats::runif(n)
+  qtruncexp(p = u, rate = rate, a = a, b = b)
+}
+
+#' @rdname distributions
+#' @export
+dtruncexp <- function(x, rate = 1, a = 0, b = Inf, log = FALSE) {
+  stopifnot(b > a)
+  ll <- log(rate) + rate*a - rate*x - log(1 - exp(-rate*(b - a)))
+  ll[x < a] <- -Inf
+  ll[x > b] <- -Inf
+
+  if(log) ll else exp(ll)
+}
+
+#' @rdname distributions
+#' @export
+ptruncexp <- function(q, rate = 1, a = 0, b = Inf, lower.tail = TRUE, log.p = FALSE) {
+  stopifnot(b > a)
+  num <- exp(-rate*(q - a)) - 1
+  den <- exp(-rate*(b - a)) - 1
+
+  p <- num/den
+  p[q > b] <- 1
+  p[q < a] <- 0
+
+  if(!lower.tail) p <- 1 - p
+  if(log.p) p <- log(p)
+  p
+}
+
+#' @rdname distributions
+#' @export
+qtruncexp <- function(p, rate = 1, a = 0, b = Inf) {
+  stopifnot(b > a, 0 <= p, p <= 1)
+  a - log(p*exp(-rate*(b - a)) - p + 1)/rate
 }